requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
click==8.1.3
colorama==0.4.4
imbalanced-learn==0.9.0
imblearn==0.0
iniconfig==1.1.1
joblib==1.1.0
marshmallow==3.15.0
marshmallow-dataclass==8.5.8
mypy-extensions==0.4.3
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
typing-extensions==4.2.0
typing-inspect==0.7.1
